sn-gateway/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.11
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
idna==3.10
protobuf==5.29.3
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
setuptools==75.8.2
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
pydantic[email]
